refactor: update component imports to use new ui directory structure
