+++ src/app/[lng]/not-found.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image";
 import Link from "next/link";
-import Footer from "@/components/footer";
+import Footer from "@/components/ui/footer";
 import { PageProps } from "@/types";
 import { useTranslation } from "../i18n";
